Move testes para a pasta correta
