+++ tests/Unit/Services/MenuServiceTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Unit\Service;
+namespace Tests\Unit\Services;
 
 use App\Models\Menu;
 use App\Models\Submenu;
